Document parameters for kube API server
Problem: Describe parameters required to connect to the Kubernetes API
server.

Solution: Updated README.rst with the parameters the f5-k8s-controller
accepts to connect to the kube API server. Most of the parameters we
get for free from the Kubernetes API, the exception being
--running-in-cluster. Removed documentation that is not applicable from
the README.rst as well, as it was copied originally from the
lwp-controller project.

F5 issue: Resolves VEL-479