The controller should monitor Kubernetes for changes
Problem:
The controller needs to handle specific configuration changes that happen
in Kubernetes.

Solution:
Created a new eventStream module that maintains an up to date list of specific
Kubernetes objects - specifically Services and ConfigMaps, but can handle
others as well.

Updated main.go so it sets up 2 event streams - 1 for ConfigMaps and 1
for Services.

Fixes LRS-533