f5-k8s-controller should handle client ssl profile
Problem:
The f5-k8s-controller should pass the client ssl profile from the
config map to the f5_marathon_lb layer.

Solution:
Added the client ssl profile to the VirtualServerConfig struct so
that it is serialized in the format expected by f5_marathon_lb.
The SslProfile struct was added as a pointer and declared 'omitempty'
so it will only be present in the output if it exists in the config map.

Updated the appropriate unit tests to check that the profile name is
present only when it should be.

Fixes VEL-570