VEL-560 Create Documentation - f5-k8s-controller
Problem: There is no explanation on how to configure the f5-k8s-controller
(big-ip controller) now how to configure it for load balancing.

Solution: Added content describing how to install/configure the controller
in a kubernetes environment. Described the relationship between Services
and ConfigMaps. Described the content of ConfigMaps. Described need for
Services to be NodePort type.