VEL-390 f5-k8s-controller reasserts correct state if it is not running for a while
Problem: Changes to the BIG-IP will not be detected and therefore not synched
back to state defined in kubernetes.

Solution: Implement similar checkpointing as in f5-marathon-lb. Issue 18 filed
to collapse this functionality.  Kubernetes state will always be correct in the
k8s controller because it will be consuming API events. But the BIG-IP will not
send out change events, therefore, once configure continue to checkpoint config
to find changes. If configuration is out of synch update it.