Fixes #24 bigipconfigdriver.py ConfigHandler thread can fail to stop
Problem:
If the stop function is called after the stop event check and before the
_pending_reset wait the thread may hang. The problem will occur because the
check for _pending_reset will return false and fall into the wait case. The stop
notify may be lost when the thread is not in the condition wait. Before dropping
into the wait case the stop event should always be checked.

Solution:
Before falling into the condition wait check the value of the _stop variable. If
it has been set there will be no wait and no chance to lose the notify.

From: https://docs.python.org/2/library/threading.html#condition-objects
" it is a no-op if no threads are waiting."

Ran lifecycle tests with no errors.