VEL-609 Report accurate unit testing code coverage for f5-k8s-controller
Problem:
Report code coverage results from unit testing for f5-k8s-controller.

DoD:
1. Get summary number that represents total code coverage.
2. Report summary number up to gitlab. Optionally, upload any artifacts
resulting from code coverage analysis that give a deeper breakdown of
coverage data.

Solution: Parse individual coverage statements for the percent covered
and calculate the mean. Finally echo that mean for gitlab to parse out
of the build output.

Artifacts are already being produced and saved.

This change only affects golang coverage. The python utility does not
produce coverage output. This is a helper script that should be
eventually turned into a project and configured as a sidecar container
to the f5-k8s-controller or its functionality written natively in the go
language to ease the hybrid maintenance headaches.