Separated and verified BIGIP host and port
Problem:
Python sdk defaults port to 443 and appends to hostname. Hostname was
being sent to sdk as host and port together, causing 443 to be appended.
The hostname and port should be separate.

Solution:
Split hostname arg into host and port; port is now passed as a separate
arg to the python sdk to override default if specified. Added
verification checks that we are getting a valid URL.

f5-marathon-lb submodule was updated, so new dependencies were needed in
Dockerfile.

Fixes #59