Controller using 100% cpu
Problem:
The controller is sitting at a constant 100% cpu utilization on multiple
platforms: GCE, OpenStack and when ran standalone. If you specify an
invalid kubernetes server when running standalone the utilization is sub
1%. Sha 52610299

Solution:
EventStreams wrap tools/cache/reflector structs and pass through a
resyncPeriod variable. This variable is a time.Duration which represents
an exceedingly small period of time (from
golang.org/pkg/time/#Duration):
A Duration represents the elapsed time between two instants as an int64
nanosecond count. The representation limits the largest representable
duration to approximately 290 years.

The controller was using a value of 5 for the resyncPeriod putting it in
a tight loop for every stream.

In case some updates are lost over the network resyncing may be
valuable. Using the constant time.Second to ensure we only resync at a
reasonable timeframe.

Updated calls to adhere to this review note as lines were becoming
long:
https://docs.f5net.com/display/F5VELCROPROJ/go
Go does not limit line lengths, lines of extreme length can become
problematic and hard to read:
https://github.com/golang/go/wiki/CodeReviewComments#line-length. Break
long lines on parameters