VEl-672 Added support for new config file format
Problem:  The bigipconfigdriver.py script gets its configuration
from two places: a config file and an argument list, both of which
are provided by the controller.  As we add more configuration, we
need to have a consistent approach to passing this information to
the script.

Solution:  All arguments (except the config file name) have been
moved into the config file.  This allows configuration changes on
the fly which were not previously possible when the command line
arguments were used.  Note: The bigip credentials are currently
static (for now, any subsequent changes to these will be ignored).
This change also adds support for the VxLAN configuration needed
to create FDB entries in the Big-IP.