Use pre-built base image (fixes #65)
Problem:
The docker image needs to be a combination of golang and python. To mimic the
official image, this means that python needs to be compiled. This is error-prone
because the upstream keyserver sometimes reports an error. Also, it adds a long
time to the build

Solution
Use a combined golang and python image produced by another repo.