VEl-672 Pulled in latest _f5.py with VxLAN tunnel FDB support
Problem:  The bigipconfigdriver.py uses the _f5.py script from the
f5-marathon-lb submodule.  This file was updated to support the
new VxLAN tunnel feature and needs to be updated.

Solution:  Updated the submodule to the latest which includes the
VxLAN support.  Also removed the old config format which was left
in for backwards compatibility (until the _f5.py script was
updated).