When apps are scaled to zero the virtual server is removed
Problem:
f5mlb-1 - When in cluster mode and apps are scaled to zero (valid
service and configmap but no pods) the virtual server will be deleted.

The expectation is that the virtual server will remain with a pool of no
members. Openshift will fail test_e2e (f5mlb-1) because of this.

Solution:
Allow for Backend.PoolMemberPort to be 0 and still be written to the
config. A new semantic of -1 designates it should not be written. Update
the getEndpoints utility to pass back an empty slice so an empty
PoolMemberAddrs gets marshaled as [] and not null - this prevents
needing to make a change in marathon-bigip-ctlr python code.

Updated logging that helped debugging - feel it's valuable to keep.

Fixes #79