Pool members cannot have unique port numbers when configured with k8s_bigip_ctlr
Problem: Due to the way services are described in the config file for
bigipconfigdriver.py, only one port can be specified for all pool
members. In kubernetes clusterIP mode, we need to loosen this
restriction so that we can use named ports where the individual
endpoints can have unique port number.

Solution: Build up the Endpoint matrix of IP:Port within the controller
code. For each address and port add to a list, this will be the
configured pool members.

Pass this list to the bigip controller formatted. Return list of
ip:port from getEndpointsForService

Fixes #85