In nodeport mode non-nodeport services are processed
Problem: When the k8s controller is configured for mode=nodeport
services that are not NodePort but have F5 resources will get processed
in error and configuration is written out.

This causes an error in _f5.py trying to process bad config.

Solution: The updateConfig boolean was moved erroneously. This boolean
MUST only be set when configuration has been changed. This means
non-NodePort services remain invalid and configuration stays protected.

Ran unit tests - systest do not catch this case
Added new unit test for this case