Go code coverage isn't calculated correctly
Problem:
The total code coverage is calculated as an average of averages per package,
not the actual percent of statements covered.

Solution:
Added the gcovmerge tool into the project to the project which helps
determine the actual statements that are covered by handling lines that
are reported covered by multiple unit tests, and using the built-in go
coverage tool to calculate the percentage of statements covered.

Fixes VEL-748