Status annotation for chosen bindAddr
Problem: A kubernetes API user may wish to know the bindAddr that the
virtual server is running on. With the ability to set the bindAddr via
annotation or configmap, it may be necessary to know which bindAddr was
picked.

Solution: The controller  sets the status.virtual-server.f5.com/ip
annotation with the chosen bindAddr for the configmap.

Also removed some useless code.