Reduce container size
Problem: Need a smaller container for the controller. The base image
used in the runtime container was a golang image, which is large and not
needed.

Solution: Use alpine as the base image, install golang in the build
container, and do not include golang in the runtime container. 
This reduces container size from 361 MB to 123 MB.

Golang installs from:
https://github.com/docker-library/golang/blob/f19fa68b57e811315e95091bb6b78c1e2f43d14f/1.7/alpine/17847.patch

https://github.com/docker-library/golang/blob/f19fa68b57e811315e95091bb6b78c1e2f43d14f/1.7/alpine/no-pic.patch

https://github.com/docker-library/golang/blob/f19fa68b57e811315e95091bb6b78c1e2f43d14f/1.7/alpine/Dockerfile

https://github.com/docker-library/golang/blob/f19fa68b57e811315e95091bb6b78c1e2f43d14f/LICENSE