Code coverage script fails intermittently
Problem: Build fails intermittently in coverage gathering.

Solution: Some tests use the increase and decrease in the number of
goroutines as indication of process running and stopping. This conflicts
with the use of defer statements. The next test may start before the
defer operation completes. This leads to incorrect calculations of the
number of goroutines and is an insufficient check.

Added determinancy to tests by removing goroutine checks and
calculations. Rely on alternate side-effects that running processes will
create due to proper operation. This affects the checks in configWriter
and nodePoller.

Linter updates as well.

Fixes #104