Upgrading F5 SDK to enable token-based authentication
Problem:
The current SDK does not support non-admin users or externally
authenticated users.  This is needed if we want the k8s BIG-IP controller
to be able to issue REST calls without needing local admin privileges.

Solution:
Upgraded the SDK/REST modules to a version that supports token-based
authentication (which is required by the BIG-IP for non-admin and
external users).