Interval timer does not behave consistently when stopped/started
Problem:
One of the unit tests (test_confighandler_reset) will occasionally fail.
It also does not detect the proper behavior of the interval timer during
retries  since the timer is turned off.  When the timer was enabled for
the test, additional issues were discovered.

Solution:
Redesigned the interval timer to use a real Timer object. Updated test
to enable the timer with a interval slightly longer than the test (to
verify it does not fire during retries, but will afterwards).

Also, rewrote the unit tests to make more use of the _f5.py common
module to verify more of the interaction between the driver and the
library.

Fixes: #124