Use commonized _f5.py and common.py from marathon-bigip-ctlr
Problem: The shared common python code _f5.py and common.py is moving to
the f5-cccl library. Before this can happen, we need to ensure all the
users (like this k8s-bigip-ctlr project) are pointing to the same
version.

Analysis: A version that works for all the current users has been
implemented in marathon-bigip-ctlr. Point at it.  Since we want to
preserve the token='tmos' feature in the master branch, add it to
bigipconfigdriver.py.