Namespace watch code crashes due to bad memory access
Problem:
Some new systests under development fail in the namespace watch code.

Solution:
Changed the calls to watchmanager.Add() and watchmanager.Remove() to
use the watchmanager object within appMgr. This particular
functionality is difficult to test in unit tests, but is tested well
with the new systests.

Fixes #160

affects-branches: master