Support for single service Ingress
Problem: Need to support Ingress resources from Kubernetes. The resources should translate to
a BIG-IP virtual server, in similar fashion to the Kubernetes ConfigMap. The Ingress resource should
support multiple annotations to be used for configuration.

Solution: Add support for the controller to watch and manage Ingress resources as virtual servers on the BIG-IP. This
supports basic single service ingress resources.