The .f5license file is missing attributions information
Problem:
More overrides are needed in .f5license.

Solution:
The uplift to client-go 2.0 included many new modules into the project,
but due to a bug in the attributions generator golang backend were not
immediately noticed. The updates to the .f5license file provide the
necessary overrides for the information that could not be auto-detected.
In addition, the update to client-go 3.0 added even more modules, and
they are also included here.

Fixes #175

affects-branches: master