Attributions generator is skipping python code
Problem:
The .gitlab-ci.yml needs to tell the attributions generator about the
requirements file in the python directory.

Solution:
Updated the command line to the backend so it knows where to look for
the requirements file for the python code. Also added the missing
license for the netaddr module.

Fixes #177

affects-branches: master, 1.1.0-beta.1, 1.0-stable