Event recording for Ingress
Problem: A user may want the state of Ingress configuration without having to view the controller logs.

Solution: Pushing Events to the Ingress resource allows users to describe the resource and view basic updates
on the status of the Ingress resource. These Events generally mirror what is seen in the controller logs.