Flatten virtual server config into resource config
Problem: Virtual Server config had a lot of nested fields, which was inefficient and caused multiple
config manipulations across the data path. It also locked one backend into each virtual server, which
will be a problem once multi-service ingress is supported.

Solution: Refactored the virtual server config to be flattened; each bigip resource is its own top level object.
The new config for each "virtual server" looks like this:
    ResourceConfig{
        Virtual
        Pools[]
        Monitors[]
     }
and the config passed to the bigipconfigdriver will look like this:
    BigIPConfig{
      Virtuals[]
      Pools[]
      Monitors[]
    }
This config is a collection of all ResourceConfigs. The way we store configurations changed from using a serviceKey to
using a resourceKey, which is a key of resourceName, resourceType, and namespace.