Modify the bigipconfigdriver.py in k8s-bigip-ctlr to work with cccl
Problem:
The cccl code has been modified to have pools and monitors at the top
level, the 'services' block has been renamed to 'virtualServers', and
'policies' has been renamed to 'l7Policies'.

Solution:
This commit changes the output format to that expected by cccl. This
required changes to the unit tests as well, as they were incorrectly
assuming ordering in the config blobs that is not a correct assumption
to make (this was exposed by the cccl changes).

Fixes VEL-948

affects-branches: master