Ingress with default path does not match to a health monitor
Problem:
When the "path" item is left out of the ingress configuration, it
cannot be matched to a health monitor.

Solution:
When the "path" item is left out of the ingress configuration it should
match all paths ("/"), but since it was empty it didn't match any. When
the path is empty, consider it as "/" so it can be matched to a health
monitor.

Fixes #206

affects-branches: master