Create multiple virtual servers for HTTP redirect
Problem: If we want to redirect HTTP to HTTPS, or support both protocols,
there needs to be a virtual server for each.

Solution: Create a virtual server for both HTTP and HTTPS if necessary. This will
be decided by the TLS secret and TLS related annotations. The TLS profile will apply
only to the HTTPS server, and the redirect policy will apply only to the HTTP server.