Sync policies when configs are created
Problem: Policy syncing across multi-service configs were only synced
when a service was deleted. However, when we grab configs to write to the BIG-IP,
we grab the first config for a multi-service Ingress and use that. If this config doesn't
have an updated policy from a different config, then we don't get the proper policy written out.

Solution: Sync policies across all configs for a multi-service Ingress to guarantee that whichever config
we use to write out the final config, it has the correct policy. This doesn't cause config updates every 30 seconds as
was originally thought.