Create pool even if pool members are gone
Problem: With the recent Ingress additions, we stopped creating Pools if
the pool members were gone. This is incorrect behavior.

Solution: Still create pools even if members are gone, and handle an empty
pool member array in the python driver. Correctly print a log message when
there are no pool members configured.

Fixes #227