Integrate Kubernetes BIG-IP Controller to use CCCL and CCCL Schema
Changes to the k8s-bigip-ctlr:
- Creation of a CCCL instance for each managed partition
- Modified create_ltm_config_kubernetes() to create CCCL-compliant schema
- Converted python unit tests

Note: This is an interim step. Currently, the creation of a BIG-IP configuration
occurs in two pieces: the golang frontend creates a config file from the
relevant configmaps and service descriptions, then the python backend performs
another conversion before passing the desired config to CCCL. The goal is to
create the CCCL-compliant BIG-IP config in one step within the golang frontend
of k8s-bigip-ctlr. Most importantly, this commit removes the dependency upon the
legacy f5-cccl/f5_cccl/_f5.py implemention.