add slack build notifications
Problem:
Travis builds did not post to repo slack channel.

Analysis:
Notifications are configured for the F5Networks namespace, meaning that pushes to branches on a pull request will send notifications, but pushes to branches pre-pull request will not. Notifications are sent on success as well as failure since there is not a "merge when build succeeds" as in GitLab.

Solution:
Generated "secure" key using travis cli tool and added notifiaction section to `.travis.yml` to post build notices to repo slack channel.