Remove default pool for Ingress and Routes
Problem: Default pools were being set on virtual servers for Ingress
and Route configuration. This would mean that if traffic for a specific host/path
didn't match any rules, it would still be forwarded to the default pool. This is
incorrect behavior. If traffic doesn't match a rule, then it shouldn't go to any pool.

Solution: Remove the default pool in the case of multi-service Ingress and Routes.

Fixes #288