Add route label watcher; set backend port properly
Problem:
1. Need to add functionality for the route-label cli argument.
2. There is a bug when configuring routes using the first service port we see,
as all route configuratiosn end up using this port, which is wrong.

Solution:
1. Add a watcher for route labels if they are provided. The label on the Route config
should be of the format 'f5type: label', where label is provided in the --route-label cli
argument. Default is to watch all routes.
 - Also removed the route-default-server-cert argument, as it is not needed.

2. Set specific expectations around configuring the backend service port, rather than just
grabbing the first we see. The expectations are:
   - If a port is provided in the Route config, use it.
   - Else if the Route is of type passthrough or reencrypt, port is set to 443
   - Else port is set to 80
This allows for explicit port configuration for every Route, rather than the ambiguous behavior
of grabbing the first port we see.