Fix SNIDefault for Routes
Problem: SNIDefault was not being set for custom profiles when configuring
https Routes. The server name was set, but a default SNI was not.

Solution: Set the default SNI for the first custom profile for a virtual server.
All subsequent profiles for that same virtual server will not be set for default SNI.