Change Secret SSL profile log to be DEBUG
Problem: The log informing users that the SSL profile name provided couldn't be
found as a Secret and instead is used as a regular path was misleading. Users may think
there was an error if a normal path is provided, even though there wasn't.

Solution: Add this log into DEBUG rather than INFO, and clear up the message.