Support Routes in multiple namespaces
Problem: Routes in multiple namespaces were not supported, since a single IP address is supplied
to the controller, and therefore would cause conflicts when attempting to create a new set of virtuals
for a new namespace.

Solution: Rename the virtuals for OpenShift Routes, and only create 2 vservers for all routes in all namespaces.