Create default client and server profiles for SNI
Problem: The controller was using the first https route's cert/key
as the default for SNI for the https virtual server. There was a bug
that either caused none of the Routes to be used, or more than one.

Upon further discussion, we concluded that we don't want to use the Routes'
certs/keys as the default SNI profile, and instead a generic or user-specified default.

Solution: Route certs/keys will no longer be used as default SNI profiles for the https
virtual server. Instead, we will allow the user to create client and server ssl profiles and
specify them via CLI. If none are specified, the controller creates a generic profile, inherited
from either the clientssl or serverssl profiles, and use those as default for SNI.