Fix copy by reference bug
Problem: When using a pre-existing config as a base for a newly processed config,
the controller would copy by reference instead of value. This resulted in the config from
the cache being updated whenever the newly processed config was updated, keeping both configs
in the same state. This prevented updates from happening.

Also found a bug from the previous data group commit that would attempt to access a record
from a list that was altered.

Solution: Use golang's builtin copy method to copy by value instead of reference. For the data group
bug, don't alter the list while iterating over it; save the values we care about and alter the list afterwards
if necessary.