Fix port update for Routes
Problem: If a port was updated on a Route object, the controller would not update the corresponding
Pool with the new port. There was also an issue that if one Route had an inactive pool, the controller
would deactivate all stored configs when attempting to equalize them.

Solution: Update an existing pool if the port changes. Also, if a pool becomes inactive, don't deactivate
all other stored configs for the Route virtual servers. Simply keep the active state for all other configs.