Deploy docs on stable branches
Problem:
The current logic in the travis file has specific knowledge about
version numbers for stable branches when deploying product docs to
clouddocs. This should be done in a version agnostic manner.

Solution:
Added logic to the docs deploy step so that it is applicable across all
stable branches.

Testing:
Ran these changes on the stable branch to see that the docs deploy step
was executed.

affects-branches: master, 1.2-stable