Client and server ssl annotations for Routes
Problem: Users may not want to use the certs/keys exposed in the Route resource
configuration. Instead, a user wants to use a pre-defined profile from the BIG-IP
to be linked to the Route.

Solution: Add annotations for Route resources to allow users to specify profiles on the BIG-IP
to use instead of creating a custom profile from the certs/keys within the Route config.