Fix DNS unit tests to be more reliable
Problem: The DNS resolver unit tests checked for specific IP addresses, but this
is not reliable, as addresses change.

Solution: Generalize the tests to expect that an IP address exists, but don't check
the actual value, as it will change. System tests will be a better way to check
IP address value.