targetPort flexibility for Routes
Problem: OpenShift users would need to hard code a Route's targetPort if the port was
something other than what we already assumed. This was an inconvenience for the user. The controller
also would not recognize port names, such as "8080-tcp", rather than port numbers. The OpenShift
GUI creates Routes in this manner.

Solution: The controller will now accept port numbers or port names in the targetPort field
in a Route config. If this field is not provided, then the controller will grab the first port
off of the service that the Route references.