Generalize VXLAN mgr; ARP support for Flannel
Problem: The current fdb record setup is named specifically for OpenShift SDN. Flannel support will
reuse the same code, so naming should be more generalized. Flannel also requires ARP entries to be
written to the BIG-IP system.

Solution: Generalize the naming of the openshift sdn package to be "vxlan". This will be used by both
OpenShift SDN and Flannel. Add a CLI argument for the flannel vxlan tunnel name. ARP entries are also
now configured and written out for Flannel.