Create resource copy for debug logging
Problem: When debug logging, customProfiles were removed from the resources struct. This could
cause future syncing issues because we removed a reference to the profiles, so our internal store
was in an incorrect state.

Solution: Don't touch the resources blob when attempting to alter what is logged (customProfiles in this case).
Instead, create a value copy of the resources into a new struct, leaving out the profiles. Log this new copied struct,
which avoids us from altering our internal store.