Set TCP profile for HTTP mode
Problem: By default, if HTTP mode is set on a virtual, the BIG-IP also
attaches a TCP profile. Since the controller does not add this profile,
the BIG-IP virtual and controller virtual never matched, resulting in
constant updates.

Solution: If the virtual mode is HTTP, attach a TCP profile in addition
to the HTTP profile. Functionality remains the same, this just allows for
consistency in the controller/BIG-IP virtual objects when updating.

Update CCCL pointer for additional virtual update fixes.