Unit tests for A/B deployment
Adding two new unit tests for A/B deployment. The first test verifies that alternate backends in Routes
create the correct number of pools and rules. The second test verifies that the data groups are correctly
populated with the weights and corresponding pool names (for each path).

Also fixed the PoolCount method to correctly count the number of pools (formerly it would count configs per
svcKey, which resulted in duplicates in some cases).