Added irule and datagroup support for A/B deployments in non-TLS environments
Problem:  Need to be able to load balance, using a ratio based
algorithm, between services that are part of an A/B deployment.

Solution: Created an irule that will fire on matching host/path
and will select a particular pool based on information stored
in a datagroup.

Restrictions:  Current implementation is only for non-TLS routes.