Sync profiles for Ingresses
Problem: Now that Ingresses are built off of existing configs, it is important that
all data is correctly copied to each new config. Profiles were not being correctly copied
when an Ingress was updated, causing some intermittent issues.

Solution: Correctly copy profiles when updating an existing config. Also ensured that configs are
written out when profiles are updated.