Bigipdriver.py hangs when exceptions occur
Problem: If the CCCL library throws a Big-IP exception (or any
exception), or if the driver encounters an exception while reading
the config file, then the driver hangs due to the thread terminating
(but not the process).

Solution: Catch the exceptions, log an error as specific as possible,
and then enter the retry loop.

Fixes: #506