Fix namespace bug when deleting profiles
Problem: When deleting unused profiles, if processing a namespace that did not contain
an Ingress that referenced a profile on a virtual, the controller would delete the profile (even
if the profile was referenced by an Ingress in a different namespace).

Solution: Add a namespace field to the ProfileRef struct, in order to tell which namespace a profile
reference comes from. This way we don't delete a profile uninentionally when processing a namespace that
the profile didn't come from.

Note: This is a temporary fix that can be properly reworked when the fix for bug 471 is in.