Fix version information in documentation
Problem:
The previous fix to extract docs version information relied on
git commands running in the containthedocs environment which
was failing and always going through the non-stable branch
path.

Solution:
Extract the docs version information in the same way as the previous
fix, but prior to running the containthedocs container. In the
docs Makefile, use these release related variables (if set) to
override the ones in the Sphinx config file.

Testing:
Ran the following commands with various values for TRAVIS_BRANCH
to test the functionality:
./build-tools/docker-docs.sh # from dev-machine
./build-tools/make-docs.sh   # from containthedocs shell

Now, browse docs/_build to ensure the generated docs have
the version string to meet the expectation as set by TRAVIS_BRANCH.