Tight loop in vxlan manager causes high CPU usage
Problem:
High CPU usage when controller should be idling with no orchestration
changes

Solution:
The vxlan manager consumer should not use a default select case for the
channel read. This allows the process to tight loop when no reads are
available. It is OK to block on this read.

Fixes #555