Fix iRule name for 11.6; datagroup record
Problem: The ssl passthrough irule in 11.6 was not validating data fields correctly. Also in 11.6, we
were passing a wildcard to a datagroup record, which needs to be escaped.

Solution: Properly format the iRule variables to be compatible with 11.6. Also, escape the wildcard string
that is passed in the https redirect data group.